Allow MultiVarTweens to work on properties that are 0.
I assume this was a bug caused by automatic conversion of FlashPunk
code, or something similar.

I haven't been able to test this properly, because haxelib doesn't seem
to like the zip files that ant creates; but I've tested it in a project
of my own. (And it's a delete-one-line change.)