Fixed:  Image Hardware Rendering not correctly on the origin.
It is possible set the origin of a Tile in a TileSheet.  This fix
replaces the original region with a new region with the correct center
when the origin is changed.

There may be an issue with the number of regions a tilesheet can have if
the origin of an image is changed enough, or a game restarts enough
times.  Also, _tileIndex may rollover.