Half Pixel Offset Not Working Right?
I was migrating an old project to HaxePunk 4.0 and noticed that the texture offsets looked weird. There was an extra pixel on the right and bottom of the sprite. Since OpenGL seems to require half pixel offsets (texel weirdness) I assumed that it was because the uvx2 and uvy2 values were adding instead of subtracting. However, when I changed it to subtraction I noticed that a half pixel on all sides were getting chopped when scaled. It seems that OpenGL isn't using half pixel offsets any more or this is being handled farther down in the HaxePunk/OpenFL code.