Fixed C++ crash on most accesses to Spritemap
The C++ target having initialized the `frame` value to 0, line 102 did not refresh the _region object, which then remained null until the frame was set again, thus resulting in a null pointer exception (somehow, not reported ?) and a silent crash everytime the _region object was used (this includes render() and get_width() / get_height()). Initializing `frame` to -1 fixed this crash.